Breaks down the hash function in BoxesRunTime by type so we can do as much as possible at compile time.  Documents various trouble points I've identified with trying to get the hashCodes aligned. Test case exercises the hashCode functions.
Deleted all the code associated with previous equality adventures.
Don't worry, I can put anything back if it turns out we have to
change course again, but for now it's noise.

Also, gives return types to the box and unbox methods which are
added to the primitive companions, so e.g. Int.box(5) now returns
a j.l.Integer instead of an Object.

git-svn-id: https://lampsvn.epfl.ch/svn-repos/scala/scala/trunk@19474 5e8d7ff9-d8ef-0310-90f0-a4852d11357a