Discovered and disproved this unlikely truth:
  scala> (1 to 1 drop 1) == (1 to 1)
  res0: Boolean = true

It was introduced in r21349 which was to fix #2535, but led to #3529.
I humbly suggest we can't afford to introduce bugs of this severity
in the pursuit of corner cases such as Ranges which use Int.MaxValue
as a boundary.  And looking at the patch I find the "after" code
a lot harder to follow.  Closes #3529.  Review by prokopec.

git-svn-id: https://lampsvn.epfl.ch/svn-repos/scala/scala/trunk@22160 5e8d7ff9-d8ef-0310-90f0-a4852d11357a