This commit is about not calling .length or .size on Lists. Sometimes it is unavoidable; not often.  I created some methods for the compiler which should probably be in the collections, which do things like test if two sequences have the same length in a more efficient manner than calling length on each.
Also, wouldn't it be wonderful to have some mechanism for finding
these issues in a more automated way.  Like say an annotation:

  LinearSeqLike {
    @badjuju("Calling length on linear seqs is O(n)")
    def length: Int = whee
  }

Or since I imagine everyone thinks they're calling length with
lists of 2 or 3, an optional runtime check which yelled if it
sees you calling length on a 600 element list, or worse:

  var i = 0
  while (i < myList.length) { // you don't want to see what this does
    f(myList(i)) ; i += 1;
  }

Uniformity of interface without uniformity of behavior leaves us
with traps.  No review.

git-svn-id: https://lampsvn.epfl.ch/svn-repos/scala/scala/trunk@23808 5e8d7ff9-d8ef-0310-90f0-a4852d11357a