This patch and the several to follow represent an attempt to break up a very large patch from martin which moves much of the compiler's typing infrastructure into the library so that we can utilize the same machinery at runtime that we do at compile time.
Said attempt was not very successful, either at maintaining working
order while committing incrementally or at subdividing the patch
into distinct cohesive units.  So let the record show that I tried.

-- some notes on the implementation --

This should not be judged as a finished work, but it's a necessary
opening step for implementing a reflection API without duplicating much
of the compiler.

The files in scala.reflect.common are destined for the library (not
the compiler, where they are now) but have not yet made the leap to
manage compatibility issues.

scala.reflect.generic is likely to be replaced by an alias to
scala.reflect.common.

The fate of scala.reflect.Code and its derived classes is yet to be
determined.

This first patch contains about the only delta against martin's patch,
which is to promote Chars into scala.reflect.  Review by odersky.

git-svn-id: https://lampsvn.epfl.ch/svn-repos/scala/scala/trunk@24973 5e8d7ff9-d8ef-0310-90f0-a4852d11357a