Added a command line script to ~/tools to count flag usages. I'm checking this in less out of the enormous demand for flag counting scripts than because I wanted to lower the barrier to people writing reusable bash scripts.
  *** DO YOU WANT TO WRITE NICE BASH SCRIPTS? ***

  ***    Look at tools/flag-usages.sh    ***

  *** It's easy to understand and full of helpful comments! ***

I'm not making any claims here about having massive bash expertise,
but I know a lot of people resist learning any of it (I was once like
you) so I wanted to lower the barrier a little. Because as a mechanism
for the composition and modification of the world of existing tools,
nothing comes close to the shell. And I know many of us write way too
many one-offs which we delete in shame and horror shortly after their
immediate purpose is served.

No review.  (I should say r-e-v-i-e-w by everyone but I'm sure
it would give me a nice pile of crucible errors.)

git-svn-id: https://lampsvn.epfl.ch/svn-repos/scala/scala/trunk@25439 5e8d7ff9-d8ef-0310-90f0-a4852d11357a